Create methods directory, move value_counts tests
