+++ python/cudf/cudf/tests/series/methods/test_describe.py
