+++ python/cudf/cudf/tests/test_series.py
@@ -5,7 +5,6 @@ import hashlib
 import operator
 import re
 from collections import OrderedDict, defaultdict
-from string import ascii_letters, digits
 
 import cupy as cp
 import numpy as np
@@ -50,150 +49,6 @@ from cudf.testing._utils import (
 )
 def ps(request):
     return request.param
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        pd.Series([], dtype="datetime64[ns]"),
-        pd.Series(pd.date_range("2010-01-01", "2010-02-01")),
-        pd.Series([None, None], dtype="datetime64[ns]"),
-    ],
-)
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("normalize", [True, False])
-@pytest.mark.parametrize("nulls", ["none", "some"])
-def test_series_datetime_value_counts(data, nulls, normalize, dropna):
-    psr = data.copy()
-    rng = np.random.default_rng(seed=0)
-    if len(data) > 0:
-        if nulls == "one":
-            p = rng.integers(0, len(data))
-            psr[p] = None
-        elif nulls == "some":
-            p = rng.integers(0, len(data), 2)
-            psr[p] = None
-
-    gsr = cudf.from_pandas(psr)
-    expected = psr.value_counts(dropna=dropna, normalize=normalize)
-    got = gsr.value_counts(dropna=dropna, normalize=normalize)
-
-    assert_eq(expected.sort_index(), got.sort_index(), check_dtype=False)
-    assert_eq(
-        expected.reset_index(drop=True),
-        got.reset_index(drop=True),
-        check_dtype=False,
-        check_index_type=True,
-    )
-
-
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("normalize", [True, False])
-@pytest.mark.parametrize("num_elements", [10, 100, 1000])
-def test_categorical_value_counts(dropna, normalize, num_elements):
-    # create categorical series
-    rng = np.random.default_rng(seed=12)
-    pd_cat = pd.Categorical(
-        pd.Series(
-            rng.choice(list(ascii_letters + digits), num_elements),
-            dtype="category",
-        )
-    )
-
-    # gdf
-    gdf = cudf.DataFrame()
-    gdf["a"] = cudf.Series.from_pandas(pd_cat)
-    gdf_value_counts = gdf["a"].value_counts(
-        dropna=dropna, normalize=normalize
-    )
-
-    # pandas
-    pdf = pd.DataFrame()
-    pdf["a"] = pd_cat
-    pdf_value_counts = pdf["a"].value_counts(
-        dropna=dropna, normalize=normalize
-    )
-
-    # verify
-    assert_eq(
-        pdf_value_counts.sort_index(),
-        gdf_value_counts.sort_index(),
-        check_dtype=False,
-        check_index_type=True,
-    )
-    assert_eq(
-        pdf_value_counts.reset_index(drop=True),
-        gdf_value_counts.reset_index(drop=True),
-        check_dtype=False,
-        check_index_type=True,
-    )
-
-
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("normalize", [True, False])
-def test_series_value_counts(dropna, normalize):
-    rng = np.random.default_rng(seed=0)
-    for size in [10**x for x in range(5)]:
-        arr = rng.integers(low=-1, high=10, size=size)
-        mask = arr != -1
-        sr = cudf.Series._from_column(
-            as_column(arr).set_mask(cudf.Series(mask)._column.as_mask())
-        )
-        sr.name = "col"
-
-        expect = (
-            sr.to_pandas()
-            .value_counts(dropna=dropna, normalize=normalize)
-            .sort_index()
-        )
-        got = sr.value_counts(dropna=dropna, normalize=normalize).sort_index()
-
-        assert_eq(expect, got, check_dtype=True, check_index_type=False)
-
-
-@pytest.mark.parametrize("bins", [1, 2, 3])
-def test_series_value_counts_bins(bins):
-    psr = pd.Series([1.0, 2.0, 2.0, 3.0, 3.0, 3.0])
-    gsr = cudf.from_pandas(psr)
-
-    expected = psr.value_counts(bins=bins)
-    got = gsr.value_counts(bins=bins)
-
-    assert_eq(expected.sort_index(), got.sort_index(), check_dtype=True)
-
-
-@pytest.mark.parametrize("bins", [1, 2, 3])
-@pytest.mark.parametrize("dropna", [True, False])
-def test_series_value_counts_bins_dropna(bins, dropna):
-    psr = pd.Series([1.0, 2.0, 2.0, 3.0, 3.0, 3.0, np.nan])
-    gsr = cudf.from_pandas(psr)
-
-    expected = psr.value_counts(bins=bins, dropna=dropna)
-    got = gsr.value_counts(bins=bins, dropna=dropna)
-
-    assert_eq(expected.sort_index(), got.sort_index(), check_dtype=True)
-
-
-@pytest.mark.parametrize("ascending", [True, False])
-@pytest.mark.parametrize("dropna", [True, False])
-@pytest.mark.parametrize("normalize", [True, False])
-def test_series_value_counts_optional_arguments(ascending, dropna, normalize):
-    psr = pd.Series([1.0, 2.0, 2.0, 3.0, 3.0, 3.0, None])
-    gsr = cudf.from_pandas(psr)
-
-    expected = psr.value_counts(
-        ascending=ascending, dropna=dropna, normalize=normalize
-    )
-    got = gsr.value_counts(
-        ascending=ascending, dropna=dropna, normalize=normalize
-    )
-
-    assert_eq(expected.sort_index(), got.sort_index(), check_dtype=True)
-    assert_eq(
-        expected.reset_index(drop=True),
-        got.reset_index(drop=True),
-        check_dtype=True,
-    )
 
 
 @pytest.mark.parametrize(
